fix(zones): ensure unset ZoneCreate i18n variable is set (#2135)
Whilst this bug doesn't present here it does elsewhere, although even then in a place that doesn't matter.

It gave a me a bit of a fright and I took a while to confirm it was having any user facing effect. Even so I figured that even though it's not a user facing bug it would be best to fix it up.

---

Uses and inline/ternary conditional which makes it very difficult to leave a variable as unset/undefined by accident. I also considered some typescript here, but seeing as this parameter value shouldn't really be here, doing it this way will make it easier to rework at a later date whilst maintaining a similar degree of safety that the inline ternary offers.

Signed-off-by: John Cowen <john.cowen@konghq.com>