feat(gateways): add new gateway views (#2158)
Add a new Gateways tab to the meshes modules. It links to a built-in gateway list view which is a listing of MeshGateway objects (which represent built-in gateways). The list’s entries link to a built-in gateway detail view. A secondary navigation provides a link to a new delegated gateway list view which is a listing of ServiceInsight objects of type `gateway_delegated`).

Remove both built-in and delegated ServiceInsight objects from the service list view.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>