fix: new service sub navigation not working
Fix the new service sub navigation not working on account of the refactoring of NavTabs now relying on automatic `route.children` which in this cases was an empty array. To make this work, I’ve done part of the refactoring I had planned for the time of unflagging the KUMA_GATEWAYS_UI features.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>