fix(services): external service lookup crashing in detail view (#2218)
Add an empty state when not finding the matching ExternalService to the service detail view.

Change the `size` parameter for the ExternalService lookup to 1000 to make it less likely not to find the right resource.

Change the ExternalService lookup to send a tag instead of a name parameter. The name parameter was incorrect anyway (but also not working as the Kuma API in 2.4.x doesn’t support it) as it filters by ExternalService name and not by ExternalService kuma.io/service tag. Note that the Kuma API doesn’t support the `tag` parameter in release-2.4 so this is only a forward-compatibility layer in case we want to fix this issue properly.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>