chore: add custom error handler for matcher not found router errors
Add a custom error handler to the Vue application that, by default, behaves like the standard error handling (i.e. any Vue realm errors are re-thrown). Specifically for the vue-router errors corresponding to "Matcher not found" errors, it adds a custom error message to the `Error` object because, for some reason, vue-router strips this out in production builds and makes troubleshooting unnecessarily hard.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>