refactor: switch package manager to pnpm
Replace yarn v1 with pnpm v8. Two options are applied in the .npmrc file: `save-prefix=''` as a one-to-one replacement for Yarn’s `--add.exact true` and `node-linker=hoisted` as a replacement for Yarn’s module hoisting behavior (which we need in workspace setups for the moment). Additionally, without `node-linker=hoisted`, `vue-tsc` breaks and reports incorrect TypeScript errors.

Remove the `deps` script. From testing the CI integration of pnpm with deleted caches, it doesn’t seem to be necessary.

Remove the unused package `@types/marked`.

Signed-off-by: Philipp Rudloff <philipp.rudloff@konghq.com>