refactor(dataplanes): moves dataplane viz to pick inbounds based on listeners (#2276)
This PR changes the way we 'find' inbounds to use IP addresses and then the fallout from that change (which means everything is far simpler)

### Finding inbounds by listeners/IP

Envoy stats is dot delimited line format but, in our case at least, it can unfortunately also include IPv4 addresses which are also dot delimited, furthermore we can't guarantee we are using IPv4 or even that the 'address thing' we are parsing isn't something like `localhost`.

We can mostly avoid dealing with any of this by using the cluster name and look in `http./tcp.` to figure inbounds out (the cluster name for the inbound is always `localhost`). But in the end doing that has meant adding significant workarounds and even then we found cases where we absolutely had to look in `listener.` (but we were able to continue to use the cluster name to pick the correct line.

Overall I figured dealing with the IP address thing and just looking in `listener.` would mean we could remove a lot of the workarounds, plus it it makes much more sense. "Look in `listener.` for the things that are listening, aka the inbounds"

This new approach does a switcharoo of IP address dot delimiters for `~` in our initial "line to JSON" pass. This lets us reliably use dot delimiters to parse the lines to JSON. We then switcharoo the `~`s back to `.` before we finally return the JSON. Really straightforwards, and I kinda wish I'd just done that originally :shrug:


### Other

Following on from the above:

- We now pick out inbound lines based on IP address rather than cluster name.
- I simplified everything and removed the majority of the additional types I'd added that felt like a hinderance rather than a help.
- Removed all the forking based on `gateway`/`data` that we had in the nested routes. Both gateways and sidecars have fully fledged 'inbounds' whether those are gathered from the dataplane overview API or from us constructing those from the envoy stats, they are both 'inbounds' and we should try treat them exactly the same (this should help with the work ongoing in https://github.com/kumahq/kuma-gui/pull/2277).

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>