chore: moves connection stats/clusters to use DataLoader/DataCollection (#2304)
Moves stats/clusters  to use `DataLoader` and `DataCollection`. DataCollection has an "auto empty state", so if no matching lines are found, then instead of getting an empty code editor we can an empty state instead.

I also changed the URI for the DataLoaders to use the `.../stats/ip-address` here so things synced up correctly. This also meant that I had to re-nest some `template`s in `DataPlaneDetailView`. Therefore it looks like there is loads of change in that file when actually its pretty much just a re-nest, there is one tweak in that file but I'll inline comment to point that out.

Signed-off-by: John Cowen <john.cowen@konghq.com>