fix(tests): unflake built-in gateway policy test (#2318)
I've spotted a couple of new flakes lately, this test one of them.

The mock here mocked out a HTTP endpoint that didn't seem to be being called at all. So I've removed that and added some `_COUNT` env vars that should always give us the data we are looking for the test.

This seems to stabilise this test, but there are still a couple of others that have popped up. Hopefully I get figure those out after this one which seems to be the most flakey.

Signed-off-by: John Cowen <john.cowen@konghq.com>