chore(dataplanes): various dataplane viz additions (#2315)
### Outbound service links

Outbound card now have a link to their respective service so you can click through to a list of dataplanes of the outbound and then click into the viz for the outbound (which is shows you it as an inbound). This means you can keep clicking through all hops, from start to finish. We infer the service name by removing the `-hash` from the end of the outbound (if that exists)

### Outbound (currently ungrouped) rules display 

Adds outbound rules to the drawer. We filter out MeshHTTP/TCPRoutes from here, and then do simple simple filtering based on the `matchers` to figure out whether we should show certain rules or not.

There are a few small fixes to the data re-shaping here and I also changed the random mocks so they link together better (across `_overview`, `stats` and `_rules`).

Lastly I added a bunch of static fixtures so its easier to test against something realistic for this. It's far too easy to mock out random data that might not actually happen. On top of these I added some straight forward e2e tests to give a bit of confidence that things are continuing to display properly.

---

- Part of https://github.com/kumahq/kuma-gui/issues/2124
- Closes https://github.com/kumahq/kuma-gui/issues/2251

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>