fix(services): fixes internal service navigation test (#2324)
Fixes (🤞 ) flakey service insight by mocking the serviceType to be `internal`

---

I found that the above change still produced the flake, then I also remembered that we also have a very similar flake in `policies / index` which made me look for commonalities that might cause both tests to occasionally flake, which points to KTable.

After looking at KTable's source I've tried adding an additional guard in our own `AppCollection` which should avoid any potentially incorrect and/or racey behaviour in KTable. There might not be any, but we won't know if this fixes both flakes or not for a little while I suppose. Thought it was worth a try.

Signed-off-by: John Cowen <john.cowen@konghq.com>