chore: hoist mesh data loading to avoid showing tabs on 404 (#2346)
I've been working on several pieces of work that cross into our Empty and Error states and I noticed that we shouldn't really be showing the clickable tabs for a mesh if we can't load the mesh in the first place, for example in the case of a 404. So, this PR moves the data loading of mesh to the same route as the tabs (and switches to use a DataLoader). MeshInsights are no longer loaded if you land on any other tab apart from `[Overview]`.

I've been considering moving all our top-level dataloading into this level of route anyway, so this was the final reason why I thought it was a good reason to do it, and I'll probably do similar in other areas following this.

Signed-off-by: John Cowen <john.cowen@konghq.com>