chore(policies): use display-name label for policy listing (#2375)
Use `display-name` and `namespace` labels for policy listings

---

I noticed that the mock for the affected dataplanes endpoint was static, so I changed that up but also that this endpoint is paged, but we don't offer any paging and we use frontend searching on the single page. I stopped short of doing anything towards changing this tho. Not sure if its on purpose or just an accidental omission? We might want to follow up with changing that soon.

---------

Signed-off-by: John Cowen <john.cowen@konghq.com>
Co-authored-by: Philipp Rudloff <philipprudloff@fastmail.com>