chore(dataplanes): revisit dpp column ordering and style (#2373)
Continuation of https://github.com/kumahq/kuma-gui/pull/2348

> I'm basing the conditional on the first item in the list having a namespace, and I'm guessing there are cases where its possible that the first item doesn't have a namespace and others do. I considered using a filter/find to check all items in your page of results, but then the column could appear and disappear when you click between pages. I'm not sure how edge case-y this is so thought I'd mention here before overthinking it.

We decided to always show the namespace column and use iconography/tooltip for the proxy type:

Closes https://github.com/kumahq/kuma-gui/issues/2370

---

There's a larger thing beginning here also with `XIcon` which always has an embedded Tooltip (conditionally enabled depending on the default slot). Thinking being, the majority of the icons we use should really have some sort of tootlip with them. This makes it really easy to add that tooltip, meaning we are more likely to add them.

Following on from this I did enough work to make sure we can replace/support our current `WarningIcon`, and I checked it where we use it in Zones only, but this will need rolling out separately (I can also do it here if its better, it shouldn't be too much work)


---------

Signed-off-by: John Cowen <john.cowen@konghq.com>