fix: center aligns text of DefinitionCard titles (#2404)
Note: I almost did this just in the `ControlPlaneStatus` component which is where the problem arises. But it seemed like if was fine to do at an application level. If anyone is unsure as to whether this is fine or not, I'd be happy to move it back to fixing it in the more focussed option within `ControlPlaneStatus`.

---

As a sidenote I noticed that all of the text rendered by `DefinitionCard` is rendered within `div`s. We should probably change that at some point to be more semantic (either using a `dl`, or `h*`/`p` elements) as the `<div>` element should be used only when no other semantic element is appropriate.


Signed-off-by: John Cowen <john.cowen@konghq.com>