chore(rules): moves rules related things into the rules module (#2347)
We've recently split off some things into a `@/app/rules` module, so I moved the Rules related `/data` and `sources.ts` into this module. I'm also looking to refactor and combine as much as possible to do with listing of our rules (we currently use an AccordionList to do that), so I've moved and reduced some things around a little to make it easier to see where I can do that.

Lastly, once i'd done that reduction/move around, I also spotted I could also close https://github.com/kumahq/kuma-gui/issues/2302 pretty easily, so I've done that here also. Due to this I also reorganized the tests around this to make them easier to read i.e. reduced the mocking required, hoisted the CSS selectors to apply to the entire file, reordered the tests to put tests of the same type together etc. The coverage is/should be the same.

Signed-off-by: John Cowen <john.cowen@konghq.com>