refactor(control-planes): refactors cp-details to use DataLoader (#2450)
First stage refactor of the control-plane-detail page (usually our landing page) to use DataLoader/XAction and some of the newer patterns we have in the application:

1. Uses DataLoader instead of DataSource (I think there is a bit of extra detail i need to look into here that might produce a separate PR, I'll detail in that PR if so)
2. Uses XAction over RouterLink
3. Adds DataLoader#loadable slot. This state is `loading` and `loaded`. I'm thinking we can use this instead of the `loading=false` property on `DataLoader`. I'm pretty sure it will allow us to eventually avoid having to use `data!` sometimes, in that the default slot can export `T` and the loadable slot can export `T | undefined`, meaning in a large amount of cases when not using the loadable slot we can use `data` instead of `data!` (also related https://github.com/kumahq/kuma-gui/issues/2432)

There's more work to come here I think, but I'm going to try and use this view to do several under-the-hood upgrades. Just an extra reminder that there are some additional complications involving `ControlPlaneStatus` (ideally I would like have its KCard in the `*View.vue` but I don't think that is possible yet)

One thing I found is that I almost changed the lists back to use the loading spinner instead of the KTable skeletons by accident due to the component usage hiding what was going on. It's highly likely I'll delete the components used here seeing as they are just AppCollection wrappers. I would much rather it was clear what was happening all in one place and I'm pretty sure these components are not multi-use.

Signed-off-by: John Cowen <john.cowen@konghq.com>