chore: Adds `uri` for generating typed string URIs for optional use with DataSource/Loader (#2453)
Adds `RouteView::uri` for generating `toString`-able typed URIs (i.e. a string but not a string).

See https://github.com/kumahq/kuma-gui/issues/2432 for more information. 

- `DataSource#default` should emit `data: Data | undefined`
- `DataLoader#default` should emit `data: Data`
- `DataLoader#loadable` should emit `data: Data | undefined`

I've tried hard to try to:

- make sure that the entire data structure of the data is presented in the type hint rather than just the name of the alias to the structure
- ensure that you can still use a bare string with DataSources, but you don't get the types (you can continue to use our current approach), partly for backwards compat during rollout but partly incase there is something I've not thought of i.e. we have an escape hatch.

One thing I wanted to avoid but couldn't was providing `sources` as the first runtime argument to `uri`. It looks like you can't mix required generic parameters and inferred generic parameters in the same call with Typescript. I toyed with putting it as the last argument (so we can later easily drop it when we no longer need it), but figured you would probably naturally want this as the first argument, so you can then auto complete the second argument.

I think the other thing its worthwhile thinking about whilst reviewing this PR is that its super important that the user provides the same cache key/identifier/URI for different DataSources in different areas of the UI and this is why we go a little bit around the houses here to stick to the "you must provide a string URI to `:src` just like an `<img src="" />`" concept.

Signed-off-by: John Cowen <john.cowen@konghq.com>