chore(deps): move back to using semver ranges rather than exact pinning (#2544)
A while back we had a little issue one dependency, that from what I remember wasn't necessarily down to dependency management but more a bug with the dependency itself (if anyone is interested I can find the trail of comments around this)

So, it was decided that exact pinning dependencies would solve the issue and we did so back in https://github.com/kumahq/kuma-gui/pull/1555.

Since then we have had lots of little issues down to duplicate dependencies and issues down to the false guarantee that we are exact pinning all our dependencies (leaving all the sub dependencies that we don't control range pinned). At this point we've kind of proven that exact pinning is giving us no guarantee we install that exact version whatsoever.

This PR reverts to the standard practice of range pinning our dependencies, meaning:

- our package manager can decide which version to install for the sake of only installing a single version of a dependency based on the constraints we give it.
- we follow what everyone else in the JS community does, meaning less surprises
- we can still exact pin individual dependencies if required (which I don't think helps in our case) or even use `resolutions` or similar for the times when we absolutely want to tell the entire dependency tree that we one an exact version of something to be used.
- we don't have to commit a `.npmrc` file (that can contain secrets) to the repository
- probably more things I could say but thats probably enough for the moment

I'd like to iteratively see where this takes us. _IF_ we have issues due to a certain dependency we should consider pinning that single dependency rather than blanket pinning the entire project.

Final Note:

I've left kongponents pinned for the moment. The current kongponents pre-releases (that we don't want) are published to `npm` with a higher semantic version that the released version, and we don't want to consume those higher pre-releases. We already ignore kongponents with dependabot config.

Signed-off-by: John Cowen <john.cowen@konghq.com>