fix: Ensure route.active doesn't force a page re-render (#2581)
Back in https://github.com/kumahq/kuma-gui/pull/2192 we made `route.children` and `route.active()` to make it easier to create sub-child navigation in parent routes (mainly our Tabs). A function was chosen as the interface for `route.active()` as the `active()` function never changes and therefore doesn't trigger a template re-render.

This was then changed to use a property rather than a function in https://github.com/kumahq/kuma-gui/pull/2186/files#r1499072744. but as this is a computed property that depends on the `route.name`, every time the routing changes the entire routing tree is re-rendered (despite what the comment in the above link says)

---

This PR changes `route.active` back to be a function, and also renames it something which describes what it does a little better i.e. I called it `route.child()` a route (at least in our application) can only have one active `child`, this naming also works well with `route.children`.

Whilst doing this I also realised a much better way of figuring out the active child of each parent route: `route.matches` gives you a hierarchical list of all the 'active' routes in the chain 'vertically'. Therefore given a single `route` and its `children`, if a child exists in the hierarchical chain, that means it is one of the active ones, and therefore the `route.child` of the `route.children` for a single route.


Signed-off-by: John Cowen <john.cowen@konghq.com>