chore: remove "Go to details" add action dropdown to all rows (#2725)
- Removes "Go to details" links from all tables
- Adds end of row dropdown with "View details" item in it

This PR resolves several issues with the "Go to details" links such as when they were shown _as well as_ a dropdown, wrapping issues, and alignment issues.

I made some amends and reused `XActionGroup` for the dropdown menu as what we are building here is a "group of actions", it just so happens this group of actions is shown as a dropdown menu. If you add `expanded="true"` to the `XActionGroup` it will be shown as a row of buttons instead of a dropdown i.e. same component for showing groups of actions with an attribute to allow you to configure these actions are shown.

I made quite a few changes that meant some `data-testid` attributes where removed, so there were some changes to tests here to reflect that (plus having to open up the new dropdown), unfortunately it also meant a tonne of formatting changes 😬 

I made a few other changes while I was there such as converting some `RouterLinks` to `XActions` (there is a long running task to do this and eventually remove `RouterLink` entirely)

---


Closes some of https://github.com/kumahq/kuma-gui/issues/2668

Signed-off-by: John Cowen <john.cowen@konghq.com>