fix(dataplanes): ensure search parameter is passed to HTTP call (#2788)
Back in https://github.com/kumahq/kuma-gui/pull/2749 we changed the remaining DataSource's to use DataLoader and `uri` (`uri`s benefit is it carries through all the types for the resource identifier/data).

Unfortunately we mistakenly omitted a `search` parameter (these params are optional). This PR adds it back in.

Double-y unfortunate is that I thought we had e2e tests on our search, which we do, but for another page (dataplane searching within a service detail) 🤦 .

https://github.com/kumahq/kuma-gui/blob/924c7a597ce785fa5d5cfdef48195181fa02f027/features/mesh/services/Item.feature#L58-L71

I copied the above test over to the DataPlane listing page so we now test this in both places. You'll see that new test in this changeset.

Signed-off-by: John Cowen <john.cowen@konghq.com>