fix: prevent a re-render on `me` submission (#2810)
DataSink began with a single static 'exported' property called `submit`.

Since then we've added several 'exported' properties to DataSink as we've built it out to cover several usecases (`data`, `error`, `writing`, `payload`)

The issue is, these properties are dynamic and will change, and even though we purposefully don't reference these dynamic properties in a DataSink consuming template, the very fact that they exist under-the-hood fires a re-render of DataSink's child content.

---

There is one place where we currently don't want a re-render to occur, that is from within `RouteView` when using the `me` DataSink. This is the code we are changing here.

This PR saves off a reference to `submit` _only on mount_ using a Vue template ref. This means we can use DataSink as an inline component/element and it has no child content to re-render.

---

Note: There is a way to do this with Vue's `defineExpose` and then avoid the extra template tag. But I found it very awkward to use along with Typescript. Vue's docs point to a further package to depend on/install in order to get things working correctly with Typescript, and even then it was troublesome. The approach I've used here works fine alongside Typescript and therefore is the favoured approach unless I come to find a better way in the future (at which point we can change this up)

Further note: I've since seen that https://github.com/kumahq/kuma-gui/pull/2811 is probably a better fix for this, but I'd still like to merge this PR seeing as its possibly valuable in areas where we don't use KTable.


Signed-off-by: John Cowen <john.cowen@konghq.com>