feat(dataplanes): disable linking of new Mesh*Services (#2853)
During parsing of the Envoy stats responses, if we find a clustername with `_msvc_` (or related), we mark that cluster/outbound as having a `kind` of `MeshService`.

Later on in the view, we can decide whether to add a link to the cards service depending on its `kind`.

Currently we only link cards with empty kinds. i.e. if its not a MeshService, MeshMultiZoneService or a MeshExternalService we don't link the card.

See linked issue for more details https://github.com/kumahq/kuma-gui/issues/2848

Signed-off-by: John Cowen <john.cowen@konghq.com>