chore: Reorders RouteView to have a root node (#2896)
Shuffles around the components in `RouteView` to ensure we have a proper DOM node as root for if we ever add attributes to the component (`<RouteView data-testid="">`)

Whilst previously `DataSource` was the single root node, `DataSource` itself does not have a wrapping root DOM component. This PR moves the root node here to be a `div` i.e. a proper DOM node.

Signed-off-by: John Cowen <john.cowen@konghq.com>