fix(onboarding): fixes empty state text and flakey onboarding test (#3058)
We had a flakey onboarding e2e test, and when I started looking into it
I saw another little issue, I fixed that in the end didn't seem related
to the flake:

- I noticed that the even when the response was empty, the GUI would say
`The following data plane proxies (DPPs) are connected to the control
plane:` and show `0` instead of saying `Waiting for DPPs`. This was down
to `.some` doing the opposite of what you expect when the array its
acting on is empty. There is a small fix in the application code for
this.
- I believe the flake was due to the test checking for the
`$state-waiting` potentially after the initial request has
responded/loaded, meaning it will never exist. As the request could
respond either before or after the `$state-waiting` assertion, this
causes the test to be flakey. I've changed the test to first always
respond with zero dataplanes until we assert that `$state-waiting`
exists, and then carry out the rest of the tests as we did before.

Signed-off-by: John Cowen <john.cowen@konghq.com>