fix(dataplanes): Moves dataplanes to use state instead of health (#3078)
Ideally this PR wouldn't be so big but! This PR does two things:

1. Switches our `online` / `offline` badges to use `.state` instead of
`.health.ready`
2. Whilst looking at this I considered checking both properties, and I
felt it was a good time to split the `data-planes/data/index.ts` into
separate files so things were easier to look at. The majority of this PR
is just the file moves. I've commented in the inline PR notes where the
more important logic changes are.

Signed-off-by: John Cowen <john.cowen@konghq.com>