chore: prevents unwanted click events firing (#3048)
This PR does two things related to unwanted "synthetic" click events
happening in our table rows. Our table rows use a synthetic click event
to simulate a HTML anchor click because we use tables for navigation
lists and we are unable to wrap anchors around table rows (`<a />`s
around `<tr />`s)

The fixes:

1. Our summary drawer has an `onClickOutside` event so that if its open
and you click outside the drawer, it then closes. Thing is, if you click
outside the drawer on a different table row, we don't want the drawer to
close, but instead change to the summary of the row you have clicked.
The `.preventDefault()` related code here prevents the drawer from
closing if you click a different row.
2. Currently if you select a piece of text that is in a table row
(probably in order to copy it to your clipboard), when you let go of
your mouse to finish the selection, it immediately fires the mouseup of
a click event, meaning you are then navigated elsewhere, which is
incredibly frustrating and has probably led to people wanting more and
more copy buttons. The `window.getSelection()` code here prevents the
navigation happening if you are selecting text.

Signed-off-by: John Cowen <john.cowen@konghq.com>