chore: moves several components to our x namespace (#3088)
We have various reasons to wrap, fix or otherwise rename some
KComponents. This means that in order to know whether to use a `X` or
`K` component, you need to know whether some sort of extra
functionality/fixing/wrapping is required.

It would be much easier if all our generic GUI components were prefixed
with `X` (we don't want to use `K` because we don't own that prefix)

This PR does several renames:

1. KBadge > XBadge (we have a spearate XBadge implem.)
2. KAlert > XAlert (just a rename for consistency)
3. KPop > XPop (just a rename for consistency)
4. KInputSwitch > XInputSwitch (just a rename for consistency)
5. CodeBlock > XCodeBlock (we have a separate implem. to KCodeBlock)

One of the consequences of moving CodeBlock to `x` is that the
`code-block` folder also contained an additional component
`ResourceCodeBlock`. I have a long running task of reducing the amount
of CodeBlock components we have (I think we had 4 or 5 at one point),
and at some point soon I plan on merging this `ResourceCodeBlock` into
`XCodeBlock` meaning we will only have 1 CodeBlock component in our
application.

Lastly, there are still some outstanding `K` components that need a
rename of a wrap, but I didn't want to do too much in a single PR.

Signed-off-by: John Cowen <john.cowen@konghq.com>